% Generated by roxygen2 (4.0.1): do not edit by hand
\name{closure_total_fair}
\alias{closure_total_fair}
\title{Total Closure of Adjacency Matrix [Fair Totalization]}
\usage{
closure_total_fair(B)
}
\arguments{
\item{B}{object of class \code{igraph} or a square
0-1 matrix of class \code{Matrix} or \code{matrix}}
}
\value{
object of class \code{Matrix}
}
\description{
Fair totalization: for each pair (x,y) s.t.
not xRy and not xRy let from now on xRy and yRx
}
\details{
If you want a total preorder, call \code{\link{closure_transitive}}.
}
\seealso{
Other binary_relations: \code{\link{closure_transitive}};
  \code{\link{de_transitive}};
  \code{\link{get_equivalence_classes}};
  \code{\link{get_incomparable_pairs}};
  \code{\link{get_independent_sets}};
  \code{\link{is_reflexive}}; \code{\link{is_total}};
  \code{\link{is_transitive}}; \code{\link{pord_weakdom}};
  \code{\link{rel_graph}}
}

