\name{de_transitive}
\alias{de_transitive}
\title{De-transitivitize Graph}
\usage{
  de_transitive(B)
}
\arguments{
  \item{B}{object of class \code{igraph} or a square 0-1
  matrix of class \code{Matrix} or \code{matrix}}
}
\value{
  object of class \code{Matrix}
}
\description{
  Useful for draving Hasse diagrams.
}
\seealso{
  Other binary_relations: \code{\link{closure_total_fair}},
  \code{\link{closure_transitive}},
  \code{\link{get_equivalence_classes}},
  \code{\link{get_incomparable_pairs}},
  \code{\link{get_independent_sets}},
  \code{\link{is_reflexive}}, \code{\link{is_total}},
  \code{\link{is_transitive}}, \code{\link{pord_weakdom}},
  \code{\link{rel_graph}}
}

