% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_incomparable_pairs}
\alias{get_incomparable_pairs}
\title{Get Incomparable Pairs in an Adjacency Matrix}
\usage{
get_incomparable_pairs(B)
}
\arguments{
\item{B}{object of class \code{igraph} or a square
0-1 matrix of class \code{Matrix} or \code{matrix}}
}
\value{
integer matrix with two columns (indices of incomparable elements,
not that these are pairs, and not sets: you'll get (i,j) and (j,i))
}
\description{
A pair (x,y) is incomparable iff
not xRy and not xRy
}
\details{
See also \code{\link{get_independent_sets}} of how to generate
all maximal independent sets.
}
\seealso{
Other binary_relations: \code{\link{closure_total_fair}};
  \code{\link{closure_transitive}};
  \code{\link{de_transitive}};
  \code{\link{get_equivalence_classes}};
  \code{\link{get_independent_sets}};
  \code{\link{is_reflexive}}; \code{\link{is_total}};
  \code{\link{is_transitive}}; \code{\link{pord_weakdom}};
  \code{\link{rel_graph}}
}

