% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rel_graph}
\alias{rel_graph}
\title{Create Adjacency Matrix of Given Binary Relation}
\usage{
rel_graph(x, pord, ...)
}
\arguments{
\item{x}{list with elements to compare, preferrably named}

\item{pord}{function with 2 arguments, returning boolean value}

\item{...}{additional arguments passed to \code{pord}}
}
\value{
square 0-1 Matrix (of class \code{Matrix})
}
\description{
Note that adjacency matrix  can also be conceived as a directed graph (DAG).
ret[i,j] iff i <= j
}
\seealso{
Other binary_relations: \code{\link{closure_total_fair}};
  \code{\link{closure_transitive}};
  \code{\link{de_transitive}};
  \code{\link{get_equivalence_classes}};
  \code{\link{get_incomparable_pairs}};
  \code{\link{get_independent_sets}};
  \code{\link{is_reflexive}}; \code{\link{is_total}};
  \code{\link{is_transitive}}; \code{\link{pord_weakdom}}
}

